+++ NEWS.md
@@ -70,6 +70,8 @@ TODO: sort out what happens to `glance.aov()`
 
 ## New tidiers, features and bugfixes
 
+- Added tidier for `summary.manova` (#729)
+
 - Added tidier for `epiR::epi.2by2` (#711)
 
 - Make `.fitted` values respect `type.predict` argument of `augment.clm()`. (#617)