+++ NEWS.md
@@ -72,6 +72,8 @@ TODO: sort out what happens to `glance.aov()`
 
 ## New tidiers, features and bugfixes
 
+- Added tidier for `summary.manova` (#729)
+
 - Added `tidy()` and `glance()` methods for `speedglm` objects from the `speedglm` package
 
 - Added tidier for `epiR::epi.2by2` (#711)
@@ -100,6 +102,7 @@ TODO: sort out what happens to `glance.aov()`
 - Add feature for glance.biglm to return df.residual
 
 - Patch bug in glance.lavaan (#577)
+- Added tidiers for `lmrob` and `glmrob` objects from the `robustbase` package (#205, #505).
 
 - Added method `tidy.systemfit()` to tidy `systemfit` class models (by @jaspercooper)
 
@@ -146,6 +149,7 @@ TODO: sort out what happens to `glance.aov()`
 - Bug fix for `tidy.mlm()` when passed `quick = TRUE` (#539 by @MatthieuStigler)
 - Bug fix for `tidy.polr()` when passed `conf.int = TRUE` (#498)
 - Bug fix in `glance.lavaan()` (#577)
+- Added tidiers for `lmrob` and `glmrob` objects from the `robustbase` package (#205, #505).
 
 ## Deprecations